Renamed randomrange.h for now. Think about using https://cplusplus.com/reference/random/
https://github.com/Synthux-Academy/simple-examples-touch/pull/1#discussion_r1442255957
