+++ libdaisy/simple-arpeggiator-touch/arp.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../utils.h"
+#include "../randomrange.h"
 
 namespace synthux {
