+++ libdaisy/simple-arpeggiator-touch/vox.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../utils.h"
+#include "../randomrange.h"
 #include "daisysp.h"
 
 using namespace daisysp;
